+++ src/components/FormRenderer/index.test.tsx
@@ -948,89 +948,4 @@ describe('FormRenderer', () => {
             fieldArrayTest({ ...schema, layout: 'grid' });
         });
     });
-
-    describe('Table', () => {
-        const schema = {
-            ...baseSchema,
-            fields: [
-                {
-                    component: componentTypes.TABLE,
-                    name: 'table',
-                    label: 'Table',
-                    description: 'This is a table',
-                    getRowId: (data: any) => data.id,
-                    items: [
-                        {
-                            id: 'id0000011',
-                            name: 'Order 11',
-                            createdDate: '2019-10-12',
-                        },
-                        {
-                            id: 'id0000012',
-                            name: 'Order 12',
-                            createdDate: '2019-11-12',
-                        },
-                    ],
-                    columnDefinitions: [
-                        {
-                            id: 'id',
-                            width: 200,
-                            Header: 'Id',
-                            accessor: 'id',
-                        },
-                        {
-                            id: 'name',
-                            width: 200,
-                            Header: 'Name',
-                            accessor: 'name',
-                        },
-                        {
-                            id: 'createdDate',
-                            width: 200,
-                            Header: 'Created date',
-                            accessor: 'createdDate',
-                        },
-                    ],
-                },
-            ],
-        };
-
-        it('should render Table', () => {
-            act(() => {
-                const { getByText } = render(
-                    <FormRenderer schema={schema} onSubmit={handleSubmit} onCancel={handleCancel} />
-                );
-
-                expect(getByText('Order 11')).toBeVisible();
-                expect(getByText('Order 12')).toBeVisible();
-            });
-        });
-    });
-
-    describe('Markdown Editor', () => {
-        it('should render a markdown editor', () => {
-            const schema = {
-                submitLabel: 'Save',
-                cancelLabel: 'Back',
-                fields: [
-                    {
-                        component: componentTypes.MARKDOWN_EDITOR,
-                        name: 'markdownOne',
-                        label: 'This is a markdown editor',
-                        helperText: 'Helper text provides users some guidance.',
-                        initialValue: '# I am a Markdown editor',
-                    },
-                ],
-                header: 'Markdown Editor',
-                description: 'This component allows a user to enter markdown and renders it in real-time.',
-            };
-
-            const { getByText } = render(
-                <FormRenderer schema={schema} onSubmit={handleSubmit} onCancel={handleCancel} />
-            );
-
-            expect(getByText('Markdown Editor')).toBeVisible();
-            expect(getByText('I am a Markdown editor')).toBeInTheDocument();
-        });
-    });
 });
